# assessment/openapi.yaml
openapi: 3.0.3
info:
  title: Keyloop Sales Lead Management API
  version: 1.0.0
  description: REST contract for Scenario C backend implementation.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/leads:
    get:
      summary: Lead inbox
      operationId: getLeadInbox
      parameters:
        - in: query
          name: dealershipId
          required: false
          schema:
            type: string
        - in: query
          name: leadStatus
          required: false
          schema:
            $ref: '#/components/schemas/LeadStatus'
        - in: query
          name: page
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - in: query
          name: size
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageLeadSummaryResponse'
    post:
      summary: Create a new lead
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLeadDetails'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /api/v1/leads/{leadId}:
    get:
      summary: Get lead details with activity log
      operationId: getLeadDetails
      parameters:
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLeadDetails'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/leads/{leadId}/activities:
    post:
      summary: Log follow-up activity for a lead
      operationId: logActivity
      parameters:
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogActivityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLeadDetails'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/leads/{leadId}/status:
    patch:
      summary: Update lead status
      operationId: updateLeadStatus
      parameters:
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLeadDetails'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/leads/{leadId}/activities/{activityId}/type:
    patch:
      summary: Update activity type
      operationId: updateActivityType
      parameters:
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: activityId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivityTypeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLeadDetails'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Activity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ApiResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
        errorCode:
          type: integer
          format: int32
          nullable: true
        message:
          type: string
          nullable: true

    PageResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean

    ApiResponseLeadSummaryList:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PageResponseLeadSummaryResponse'
        errorCode:
          type: integer
          format: int32
          nullable: true
        message:
          type: string
          nullable: true

    ApiResponsePageLeadSummaryResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PageResponseLeadSummaryResponse'
        errorCode:
          type: integer
          format: int32
          nullable: true
        message:
          type: string
          nullable: true

    PageResponseLeadSummaryResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/LeadSummaryResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean

    ApiResponseLeadDetails:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/LeadDetailsResponse'
        errorCode:
          type: integer
          format: int32
          nullable: true
        message:
          type: string
          nullable: true

    ValidationErrorResponse:
      type: object
      required:
        - success
        - errorCode
        - message
        - data
      properties:
        success:
          type: boolean
          example: false
        errorCode:
          type: integer
          format: int32
        message:
          type: string
          example: Validation failed
        data:
          type: object
          additionalProperties:
            type: string

    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - path
        - timestamp
      properties:
        status:
          type: integer
          format: int32
          nullable: true
        error:
          type: string
          nullable: true
        errorCode:
          type: integer
          format: int32
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time

    CreateLeadRequest:
      type: object
      required:
        - fullName
        - email
        - phoneNumber
        - vehicleOfInterest
        - source
        - dealershipId
      properties:
        fullName:
          type: string
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 120
        phoneNumber:
          type: string
          maxLength: 40
        vehicleOfInterest:
          type: string
          maxLength: 120
        source:
          type: string
          maxLength: 80
        notes:
          type: string
          maxLength: 1500
        dealershipId:
          type: string
          maxLength: 80

    LogActivityRequest:
      type: object
      required:
        - activityType
        - description
        - createdBy
      properties:
        activityType:
          $ref: '#/components/schemas/ActivityType'
        description:
          type: string
          maxLength: 1500
        createdBy:
          type: string
          maxLength: 80
        occurredAt:
          type: string
          format: date-time

    UpdateLeadStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/LeadStatus'

    UpdateActivityTypeRequest:
      type: object
      required:
        - activityType
      properties:
        activityType:
          $ref: '#/components/schemas/ActivityType'

    LeadSummaryResponse:
      type: object
      required:
        - id
        - fullName
        - email
        - phoneNumber
        - vehicleOfInterest
        - source
        - status
        - createdAt
        - updatedAt
        - activityCount
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        vehicleOfInterest:
          type: string
        source:
          type: string
        status:
          $ref: '#/components/schemas/LeadStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        activityCount:
          type: integer
          format: int32

    LeadDetailsResponse:
      type: object
      required:
        - id
        - fullName
        - email
        - phoneNumber
        - vehicleOfInterest
        - source
        - status
        - createdAt
        - updatedAt
        - activities
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        vehicleOfInterest:
          type: string
        source:
          type: string
        notes:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/LeadStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        activities:
          type: array
          items:
            $ref: '#/components/schemas/LeadActivityResponse'

    LeadActivityResponse:
      type: object
      required:
        - id
        - activityType
        - description
        - createdBy
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        activityType:
          $ref: '#/components/schemas/ActivityType'
        description:
          type: string
        createdBy:
          type: string
        occurredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    ActivityType:
      type: string
      enum:
        - CALL
        - EMAIL
        - SMS
        - MEETING
        - NOTE

    LeadStatus:
      type: string
      enum:
        - NEW
        - CONTACTED
        - QUALIFIED
        - WON
        - LOST